Put macro into its own mod.
This will enable us to export further things from
the autocxx crate itself.
